requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0
Pillow>=8.0.0
numpy>=1.21.0
matplotlib>=3.3.0
scipy>=1.7.0
scikit-learn>=0.24.0
tqdm>=4.60.0